Cast fd before comparison

+++ junction/kernel/file.cc
@@ -95,7 +95,7 @@ bool FileTable::Remove(int fd) {
   rt::SpinGuard g(&lock_);
     
   // Check if the file is present.
-  if (fd >= farr_->len) return false;
+  if (static_cast<size_t>(fd) >= farr_->len) return false;
   if (!farr_->files[fd]) return false;
 
   // Remove the file.
